fix: createApplicationSSE import 분리
